hot fix for Insert() param type err

+++ include/BPlusTree/BPlusTree.hpp
@@ -6,6 +6,6 @@ class BPlusTree {
  public:
     BPlusNode::NodePtr root;
     BPlusTree(int degree);
-    void Insert(int key, int value);
+    void Insert(long key, long value);
     void Print();
 };
